Makefile working with Example SQL C API
Through a bit of trial and error with the makefile the program now
compiles an extremely basic example use (from the mySQL website) on
using the C API for the mySQL adapter.

The ignore file is being funny. I don't know if it's because this is the
initial commit and the ignore file hasn't been commited yet, but the *.o
and *.out and obj/*.o aren't working correctly right now.

Using .gitkeep in the obj directory in order to keep the directory that
way the makefile process doesn't get confused when there is no obj
folder to place the o files into.