Add driver file. Updated README for verifying
Added instructions in readme on how to run valgrind for the program and
to use the md5sum function. Granted I haven't given a hash for a user to
verify in the README, but it's 4c935d83dc569598fa53abda7b93e380 as of
this writing.

Also moved the main function into a driver file called green-serv.c (I'm
so creative I know) that will continue to be used as my testbed for my
code as I code. Eventually it will probably handle daemonizing and
bootstrapping the server start-up process.