Scope From DB by ID
a.out md5sum:
709aae672b607257dcfc5d915d071069  a.out

Updated config file to include campaign defines to define the scope as
well as added in functions to retrieve the scope by it's id in the
database. If the scope is not define then the zero-ed version of the
struct is returned, this had '\0' for all 9 characters and a -1 for the
id.

Added some functions to make interfacing with scope structures easy and
uniform. These two functions gs_scope_setId and gs_scope_setDesc handle
zero-ing whatever values they need to before assigning, this makes sure
that strings are properly initialized before attempting to do anything
with them. It also makes the source code uniform, the convention for
model functions will bs gs_modelname_functionName. I've grown to like
prefacing anything that needs a namespace with snake-cased naming, but
leaving whatever I would normally call the function (in code that I
wouldn't worry about namespace collisions) as camel cased.

Something I'm debating right now is having database functions take the
connection as an argument, this would mean that calling code would
handle grabbing the connection and sending it out, but by calling the
mysql_library_end function in each of the sub functions I can't help but
think that I'm missing out on the performance boost from it caching
memory (the cause of the 'still reachable' mem-leak if no call to the
library end function is made)