Rearranged Directory Structure. Flag.h
a.out md5sum:
3347104d176d7999db597c79eee4abbc  a.out

Added flag.h and some rudimentary and simplistic argument parsing. I
have yet to try to give it an invalid input (although I'm curious on how
it will behave).

Rearranged directory structure to be like the one described in 779eac0f
it doesn't have the network or driver directories because I haven't
begun working on those yet. I intend on having the JSON serialization
and database interactions and models all done first. Then I will deal
with crafting HTTP response headers and polling network connections.

The main purpose of the files that will end up in the driver directory
are to handle process forking, so that when I get a response for a
resource that has nothing to do with any other resources I can spawn a
new process or use a thread pool to serve back the response without
delaying the delegating main function. I think it will be difficult to
code and debug, so I'm doing it last. Single process first. Then
multi-processing. I think that the multiprocessing will really just be
at the driver level and maybe the network level. The database
interaction shouldn't have to worry too much I don't think.

One of the things I haven't given much thought to is the directory
structuring of the header files. Do they need to be structured or
organized in some way?