Insert Comments and Comment Json
md5sum a.out
2ba3cf91b520cb99cb701f34225cb080  a.out

Inserting comments from the struct works as expected. I perform a query
for the comment object in the database instead of calculating what the
timestamp is. This is to provide consistency in the results and also if
any truncation occurs from the users submission to the response I want
them to be aware of it.

You CAN overrun the buffer and get a stack smashing warning or a
segfault if the character array passed that will eventually hold json
isn't big enough. A wise decision would be to dynamically allocate this
and increase the buffer size as neccesary, but I'm not sure on the
cleanest way to do that at this current juncture. I'm not going to worry
about it right now and will just allocate enough memory as I go.
Eventually this will probably need a refactor, but for now, sprintf will
have to do.

Updated sql to use timestamps instead of DateTime because the timestamp
can auto add to the current time, which is what's desired. Added
pagination define for macro purposes and a test of the comment being
added.

A Key thing to note here is that comments are limited to 140 characters
(141 technically because of \0) which means that anything more than that
is truncated by the strncpy and bzero-ing that occurs. This is on
purpose, and if a user doesn't like it they can cry to themselves about
how twittery things are mean.