Refactored connection to MYSQL passing
a.out md5sum
dadfc0a9bc8c190caea77f9c7d2253f1  a.out

Moved connection outside of function calls. This makes the calls to
interact with the database require a connection, and leaves it up to the
callee to close the connection and free the library to ensure no leaky
database connections. But this will be neccesary later on when multiple
processes are using the same connection.