Add getting comment by id
a.out md5sum
fc60db8586d98b2fa0cd5c3fe4d2f57d  a.out

Added function to get a comment by it's ID from the database and tested
it. Will return a zero-ed struct if the id is invalid so that's what to
check for any type of 404 response later on down the line.