Getting Pages of comments working.
a.out md5sum
9cc9b570edad760215627386a2587fc4  a.out

It is now possible to retrieve a list of comments from the database.
Safety has to be made to malloc enough room for a page of results. An
interesting trick I read (for paginiation purposes) is to request one
more than you intend on paginating with and that's how you can tell if
there is a 'next' page or not. It's pretty clever.

Something else to begin considering is the use of the memory-only tables
that I intended for use as a cache. From some quick googling it looks a
lot like InnoDB and other engines have their own cache and "warm up" as
well, so it might be silly to try to do my own. However, I could use
the cache tables to load up test data or a test environment of some
kind. I just really want to play with an implementation of database
caching for fun to be honest.

Also added a macro to stringify other defines. This is so I can use the
number of rows per page as a string within other queries.