Ancestor Query scope for comment paging
md5sum a.out
8072fb77085dd1148dd44f7d691a3d1e  a.out

Added scope id to the comment page query and get comment pages function.
This is to facilitate groupings of entities in the database by their
respective scope. Which can be used by an implementing party as a
campaign. For example, Green Up VT could occur 1 year, then the next,
and use two different ancestries to enable partioning of the data in
such a way that makes statistics easily processable.