Added 2/3 Marker Database Functions
Add functions to get pages of markers and to insert markers into the
database. No JSON function to nicely output each marker yet but that's
soon to come.

Just a side note, if you get a stack smashing error after changing a
query in a define, chances are you need to look into how much space was
allocated to hold the query you changed and make sure that it contains
enough space for the full query. Otherwise you'll end up smashing the
stack and overrunning your buffer.