Add single Marker to JSON function
md5sum a.out
f8f3036d8b0ee193f87ac0229db3f2a7  a.out

Created simple JSON representation of a marker. Note that it's not doing
anything fancy like following the comment ID to bring out the message
associated with the pin or anything like that. I'm just trying to get
basic service going for all of these things. The 'higher' level features
like having JSON with a mix of marker/comment data can come later on
after all of this is working correctly.

Note that Version 3 of the Green Up API introduces a boolean 'addressed'
field into the Marker. I am choosing to ignore this for now while I work
on this.