Updated README
Added a small amount of help text to a few sections of the readme
detailing that it's possible to check your binary's hash against one
from the git log. Also informing the user that they NEED to make a
config.h file and to use the example one as a template of defined terms
they need.