Heatmap Model Implementation
md5sum a.out
8191e56ee56845255c053bdc927a66b1  a.out

Created implementation of heatmap model. To be honest, I'm not happy
with name of the model, it's a piece of the heatmap so I want to call it
something better. Like heatzone, or heatpoint, or something. But I'm not
sure what I would call it. IntensitySpot? For now, heatmap works well.

Updated MakeFile to reflect new model file, and changed the linker to
just do *.o so I don't have to keep updating the linking portion. While
it does pay to be explicit about these things, I'm throwing everything
into one obj folder so it's not really a problem. I don't do any partial
compilation and linking either so all is good. Although it might be
interesting to try to do that.

Another thing to note is that the setId functions are all the same
really, same with setting scope Id. I don't think there's a way to do
inheritance in C (easily at least, function pointers?). But this would
be the time for it I guess. I wonder if doing it as a generic struct
with an id field and doing a cast would allow me to do it.