Heatmap JSON, Recommendations and const
md5sum a.out
9d807fd10a60540d0e01ce3637beb184  a.out

Add toJSON function for heatmap model. Works well and creates valid
JSON. Tested and doesn't leak or have problems so long as your heatmap
model is properly initialized and that is the responsibility of the
calling party anyway.

Added recommendations on character buffer size for the jsonOutput of the
toJSON family of functions. These are based off of the nearest power of
2 that is higher than however many bytes would be strictly neccesary. We
could be fancy and use sizeof(type) and some strlen action to determine
exactly how much space is needed, but estimating at the next power of
two is likely to be safe no matter architecture of your system (32/64
bit)

Since the JSON functions do not modify the incoming struct object and
only use it as a read-only object it makes sense to inform the user of
that guarantee by using the const keyword, so this has been added.

Something I could do, is place recommendations into the model header
files or the json header files and use those instead. This would keep
all of those in the same place and provide an easy access point for
implementors.