Insert Heatmap into DB
md5sum a.out
c8c7651e0ef8521a298072c77f748a93  a.out

Function to insert heatmap into the database is complete and working
correctly. It's important to note here that when calling any database
function you must check the id of the struct being returned to you for
validity. If you don't there's a good chance that if you messed
something up (like your scope_id was invalid) then you need to take a
different course of action in whatever the calling party is. So always
check for valid id's of whatever the entity you're getting back is.

Something I've been thinking about is the consolidation of heatmap
points together. It's simple to say: if lat and lon match exactly, then
merge the two together into one entity and just update the intensity.
Which is great and will work fine once I implement that.

But what I do for partial matches? For example, 44.204 and 44.20, these
are different entities but depending on the precision specified by a
calling client, may need to be merged together (a precision of 2 would
merge these two into one intense point). I think the end result that I
will implement is that an exact match updates intensity only. And leave
it to the calling party to deal with it. My select statement will have
to take precision into account.

But first, the merging needs to happen.