Heatmap merging on insert
md5sum a.out
81eaa9a2bf4b0d343c995651e3ce1717  a.out

When submitting a heatmap point for insertion, if the latitude and
longitude match then the point is updated and the intesities are merged.
Of course, it's not just latitude and longitude but scope_id as well
other wise we'd need a different database for each scope, which would
defeat the point of having it in the first place.

A bit of funny behavior from the mysql_insert_id function was causing
some issues with the returned structure at first. It turns out that
mysql_insert_id, despite the documentation, doesn't return the id of the
affected row if said row was updated by it's id in the first place. I
guess they just assume that you know the id if you just updated by it.
Which is a totally valid assumption I guess, but makes doing an
update/insert conditional just a tad more bothersome for me. One if
statement won't kill me though.