Report header file
md5sum a.out
d382cbb45726172ce8f8d6762ed6d7f3  a.out

Added function prototypes and struct definition for the reports entity.
This entity is used to keep track of client reported bugs and errors.
One of the biggest things is that the content of this structure, if we
assigned it from the database would be massive. 64Kb space is available
for storage in the database for each of these fields.

This is clearly unneccesary. But just thinking about stack traces and
how long they can be, a good amount of space should be available. So, if
GS_REPORT_MAX_LENGTH is not long enough, we can always increase it, or
we could make the content of a report entity be malloc-d and realloc-d
for space at runtime. But I'm really trying to avoid having to call free
on objects within the structures themselves.

This bears some more consideration, but since it only affects the model
level (the database will use whatever defines we have) I can leave this
as a side note for later and maybe revisit the issue should 1024 bytes
not be enough space for the full message sent.