Implementation of reports entity
md5sum a.out
31b2b8bb7a26f8a3250259a9f22fcfd7  a.out

Basic setters and initialization of a report entity. Something that
should be added to it is the ability to get the sha256 hash of a given
string. This will be next, I believe that the functionality to do so
should be available in the openssl libraries.