Report toJSON method implemented
md5sum a.out
6dd792fcee329b4f06d477f3968558dc  a.out

Report now has a JSON format defined in the helper file. Something to be
careful of here is that I've updated the setOrigin and setAuthorize
functions for reports to call the sha256 function directly. This is
because It saves the calling party the trouble of doing so and keeps the
dependencies more minimal.

An oddity that occured was that if I don't use strncpy in those two
functions I overrun the field size in the struct itself and can bleed
into the other ones. If you do sha256(auth, gsr->authorize) directorly
you might set the origin parameter to be an empty string. I wonder if
thats from the [65] = '\0' in the sha256 function. I bet you it is.