Updated sha256 to not overrun struct fields
md5sum a.out
40f386b0c537747cda034473f8c014b7  a.out

Turns out my hunch was correct (see previous commit message). It was the
sha256 button's use of 65 instead of 64 that caused the problem with
setting the memory 'next door' to the field being set to \0 that was the
problem. This has been fixed and no more indirection buffer or strncpy
is used.