Added insert report and report queries
md5sum a.out
30640c8d4c036d24a83add1be2fce3b6  a.out

Added insertion function for reports into the database and created the
constants for the queries themselves. Realized that the decimals are
wrong and that we need leading zeros in the queries and the longs for
the mantissa will need to be a bit funny. For example:
44.05 is a 44 characteristic and a 5,000,000 long. Why? Because
otherwise it wouldn't be represented correctly.