Get reports page function added
md5sum a.out
d85aca61384cf8cfa02722881b7eac64  a.out

Added function to get a full page of reports from the database. Works
well, no memory leaks or anything. With this, the model interactions are
pretty much done. There are more complicated queries that could be made
and created, such as getting comment messages associated with markers
within the same thing and returning json or something. But really, those
are a create as I need them function.

There are some json functions that need to be made, such as correctly
formatting the paginated results, or having a status or error message
response in json. The status/error functions can be implemented when I
begin work on the http layers. The Paginated bits I could do now or
later when actually dealing with them.

There's also the fact that we've added an 'addressed' field to the
markers in the main greenup application API (which is python and GAE) in
order to see if markers have been addressed or not. This is a simple
boolean that we can add in to the model. Should be interesting to see
how much actually has to change in order to support it. This will give
me an idea of how maintainable or not the current C architecture is.