Setup testing framework
md5sum a.out
58a0529c1584818bf8bb9439e2a3dc19  a.out

Setup a tests directory and a general structure in the makefile for the
tests themselves. I was motivated to do this by needing a way to test
primehunter326's adding and subtracting functions.

The adding function fails on the current test in the mantissa category,
and the subtraction works fine, but I haven't tested it on something
like 22.5 - 10.6 yet. But I will soon. I am commiting this code to show
primehunter326 then new testing framework. Also, I just want to call
attention to the fact that whenever I change the main source (non
testing framework code) I'm md5sum-ing the a.out file and including that
in the commit message for validation.