Refactor createDecimalFromString
md5sum a.out
7db38b199acf3d296e06f19a141fca2c  a.out

Made the createDecimalFromString function. This function rely's on a few
magic numbers and this commit message details the Decimal struct itself
a bit more for anyone dealing with it. So basically, if you play with
Decimal, and get confused... come read this message again or the copy of
it that I'll probably end up putting in decimal.h or decimal.c.

The Decimal structure is composed of two fields. The left and right.
Which correspond to the mathematical concepts of characteristic and
mantissa. The characteristic is easy, a simple signed field. The
mantissa is more troublesome due to the need for leading zeros.

A unsigned long value cannot store these leading zeroes as that makes
absolutely no sense. So instead, consider the table:

+-----------+---------------------+
| Mantissa	| Long Representation |
+-----------+---------------------+
|.5	   	| 50,000,000	      |
|.05	| 5,000,000	      |
|.005	| 500,000	      |
|.0005	| 50,000	      |
|.00005	| 5,000		      |
|.000005  	| 500		      |
|.0000005  	| 50		      |
|.00000005 	| 5		      |
+-----------+---------------------+

As one can see, the zeroes are essentially fliped around. The reason for
representing the mantissa in this way is wholy based on the need for
precision over the mantissa (Due to representing latitude and longitudes
where the precision does very much matter) and also because when a
Decimal structure is transformed for a database query, it is sprintf-ed
with the 08%lu formatter for the mantissa. Because of this, the simple
.5 MUST have the neccesary number of zeros on it's left side to format
correctly.