Updated createFromString function to handle zeros
md5sum a.out
09dfda27b2264f3b8ccdbf800959ce35  a.out

Updated the strategy to dealing with leading zeroes within the mantissa
of the decimal. This function isn't quite done yet. It seems to pass the
testing framework code, but it also seems to fail during the actual
testing of the program itself. Is this because of the returned value
from the database? I'm not sure, but I do intend to find out.

What I should do. Is write integration testing for each model. THAT
would be a wise idea and a good way to tackle the problem at hand.