Updated JSON lu to %08lu
md5sum a.out
dd9bb5343dd1dabeeb7b6bc6e234f102  a.out

It figures. The reason the tests for Decimal pass with flying colors and
the regular testing of the program through it's main function is
because, quite simply, I had forgotten to update the format string in
the JSON formatters that handle the mantissa of the Decimal.

This has been fixed and it's working well again.