Updated string -> Decimal function.
md5sum a.out
f01f65d71694cd03f69fcca62d928559  a.out

The mantissa appears to be working correctly now. In order to represent
values between -0.1 to -0.9 we need to have a sign bit. This is kind of
reinventing the wheel here on floating point arithmetic. But hey, that's
pretty fun to do anyway.

The magic numbers have been removed from the calculation since they
wouldn't have worked for very long anyway. And the unsigned subtraction
arithmetic has been fixed up. This is in part thanks to the user chux on
StackOverflow for answering a question I had and pointing out that
unsigned subtraction is defined as a - b = a -b + (ULONG_MAX+1) which
helped me properly define the subtraction on the mantissa.

Reference:
http://stackoverflow.com/questions/19095288/number-above-ulong-max-during-unsigned-arithmetic

Also, chux pointed out the use of 9 in the for loops in the ghetto
string copy taking place in the String -> Decimal function that we might
run over our buffer a little bit. While this is true by looking at the
small piece of code posted to  stack overflow, this is not true within
the scope of the program since we use strncpy and bzero everywhere. So
even if 8 places are taken and a 9th written without a \0, we never ask
for the 9th place of the array anyway and therefore don't risk overflow.