Comment Test
md5sum a.out
f01f65d71694cd03f69fcca62d928559  a.out

The difficult thing about the tests is that we could break it up into
extreme unit tests, and just test the comment and make sure it's
properties are all set with alerts and whatnot. But we know that works
because we've made it this way already before we started testing. So
really, this comment-test file isn't a unit test, but a
functional/integration test with the database.

Unfortunately, because it uses the database it's cc command is rather
gross and long and filled with mysql link libraries and mysql CFLAG
arguments. Not to mention that since I was lazy and wrote obj/*.o for
this linking as well, if you do make then make tests, the compilation
will fail due to multiple main functions. This isn't really that big of
a problem because for some reason the opposite isn't true. Running make
tests and then make doesn't affect the compilation of the primary
program at all.

The 'some reason' here is probably because the compilation of the main
program doesn't attempt to use the tests in anyway.