Report Test and buffer safety
md5sum a.out
10355dc7d618a2a09d763bd3035c34c1  a.out

In making this test, and getting it to not leak memory or stash smack, I
discovered an interesting bunch of stack overflow problems. Why these
didn't manifest in the green-serv.c file is beyond me, but they manifest
in this test if the json buffer isn't large enough.

The interesting behavior in the test was that the authorize and origin
fields would be overwritten by the content during the creation of the
escaped json version of the report. This would then cause the database
query to fail because part of the query would be overwritten by the
auth's bogus values that were too long.

If we want to be really careful, I could do each query formation in
seperate steps with explicit snprintf instead of sprintf. I've done this
partially in the gs_reportToJSON function to try to remedy the issue,
but really to prevent a stack overflow the JSON buffer needs to be large
enough to handle the escaped strings. So it's either dynamic allocation
or large allocations per controller / database interactions.