Added heatmap test
md5sum a.out
10355dc7d618a2a09d763bd3035c34c1  a.out

Added heatmap test code pulled from green-serv.c into it's own file. No
leaks.

I'm thinking that as a way to fix up all the issues for a stack overflow
from the content in comments and reports being too large that I'll need
to make a new branch after merging this tests one in, and then set the
JSON size for each test to be something rediculously small. Then, if I
get a buffer overflow, go in and fix things until it's all safe and no
buffers are overran by the likes of the terrible: user content.

I thought I was safe due to the truncation that occurs from the models
but finding out that the toJSON method was modifying fields of the
structure was a tell tale sign of the code not being as safe as I had
presumed it was. This will be fixed.

The last test code I want to write before merging back to master will be
the tests for the json helper.