Add safe gs_scopeToJSON function
md5sum a.out
63f7112975bd611e01d9bebf3ef933ba  a.out

Created gs_scopeNToJSON function that attempts to guarantee that the
output buffer will not be overran. Tested in scope-test.c and it seems
to work, but I won't really know for sure until I write the safe version
of gs_reportToJSON since that's the function that manages to smash the
stack anytime the output buffer is not large enough.