Create Safe report to JSON function. Update readme
md5sum a.out
62b38a456c466b9b1b1cfe9177bbd1f5  a.out

Updated the readme to handle the new JSON return flag for compilation.
If this flag is defined (and it should be defined as "return -1;") then
the safe JSON functions will not attempt to create the JSON string
itself and will instead return -1 from the function call. This is to
avoid any trashing of the stack. Even though the safe functions use
snprintf with the size passed in, if a calling party were to pass in the
wrong size then the 'safe' functions could still malfunction, so
regardless, the functions print a warning if theirs a slight chance that
the stack may have been overran.

Added safe gs_report JSON formatting function and tested it. If you mess
around with the JSON_LENGTH define in the test for the report you'll see
that you can get back partial JSON's that are not valid if you don't
have the RETURN_ON_JSON_RISK flag defined.

Fun Fact:
- This is the first time I've put actual code into a C Macro. I've been
  trying to avoid it since it seems like an easy way to write unreadable
  code, but in this case it's pretty readable and semantic.