Added safe marker JSON function
md5sum a.out
aeaa8e29670ef584593c0fa5fd5fd320  a.out

Added safe marker JSON function. It functions the same way as the other
ones, printing out to stderr if there's a problem and it doesn't know if
it's managed to print a full and valid JSON string or not.

So far, from my observations, even setting the output buffer low I'm not
trashing the stack. Of course I have to wonder what would happen if I
did something like:

char s[10];
...
gs_reportNToJSON(&r, s, 100);

But then that's really user/developer error and not the libraries fault.