Added safe comment to JSON function
md5sum a.out
4643b4e997eec1b62e9f8c7c5ba09098  a.out

Added a safe version of the gs_commentToJSON function called
gs_commentNToJSON. The N addition is something I'm stealing from how
sprintf usings snprintf to sort of depict a bit of a safer function, and
also in that it returns the number of characters written and whatnot.

With this. I am done writing the safe functions for the models which is
good. I can now move on to more 'controller'-esk code that operates on
the data in the database and uses the models I've defined.

During the course of the controller code creation I will probably have
to create new Json helper methods to print off pages and those types of
things, but all in all it will still be enjoyable. I can't wait to get a
simple heartbeat controller and network interface going. I may skip into
just doing a full stack of the heartbeat to the controller through the
network to test it.

Evan suggested function pointers today for the calling by the
controllers. I'm not sure how to do it in an elegant way, but no doubt
it will be fun no matter what.