Updated Tests directory and MakeFile
a.out
8661df45f1de51cf48af8c25233f2765  a.out

Added some simple defines into the makefile to make it a little less
wordy. I'm contemplating using some pattern matchings to make the
building a little less redundant, at least with the test directory. I
like being fairly explicit about what depends on what within the main
program, it just seems safe to do that.

Introduced the .PHONY into the makefile as well so that clean and tests
are always ran as if their dependencies require rebuilding as well.
(Since tests should be clean).

The only thing I still want to do is to use savepoints in the mysql
database during my tests so that the 'production' level database isn't
actually affected by the tests. Although this is a pretty good way of
populating the database with a lot of data.