Began Work on basic and simple network code
-- no md5sum because no changes have been done --

Wrote a few helper functions to create http responses and setup some
properties of structures. Committing now because the current net.c, if
compiled with it's test function renamed to main, will listen for a
single connection on port 80, which you can hit by simply doing
localhost in the browser or in POSTMan (if you have that installed)

Valgrinded to make sure it works, it does. But I want to switch to using
fd sets and select to make async code easier later on. So I'm commiting
now.