More networking basics
-- no md5sum --

Got a ragtag http parsing function to populate a simple http request
struct that stores the information I care about. I think it could
probably be made simple if I made the assumption that incoming requests
would be valid http all the time. But that seems a dangerous assumption
to make.

It is working right now though with no real trouble which is rather
nice. If I can abstract the core pieces of this code into their own
functions then I might get the beginning of some network code the main
driver function can use to determine controller calling.

for example, with a GET and a url of "","/", or "/api" I could redirect
to the heartbeat.