Updated test_network to take function pointer
with the function pointer in place, I can have any 'callback' function
so long as it takes my threadData struct as a parameter and returns
NULL. This might be useful during testing.

One of things about using the detached threads is that one has to be
careful when exciting the main thread of execution, as the other threads
depend on it to live. So a simple sleep at the end allows for enough
time for the other threads to finish their processing and terminate. If
the threads do not terminate then valgrind will report that memory was
loss due to the threads not being properly deallocated.