Detached and Unattached macros for threads
-- no md5sum --

Added define to net.h of DETACHED_THREADS, that if defined causes the
network test function (and eventually the network functions themselves)
to run using detached threads. It is _safer_ to have the threads
attached and joined before allowing any to overwrite each other. That's
my gut feeling.

It may be the case that after I detach it, I could redeclare that thread
anyway without needing to keep a reference to it. But still. Eventually,
a thread pool may proove useful for amortizing the performance of the
network/worker threads. It's supposedly expensive to fire up bunches of
short lived threads (I always thought it was crap to do it with forks
that way and that's why people recommended threads, but I've been wrong
before)