Added Network To main project
md5sum a.out
74d7d6401ff84d2b0b2fe493a600dbfc  a.out

Added network and routing to main project. Note that the routing isn't
working correctly yet, but the network does fire up when you start
green-serv's a.out now. I need to do unit testing against the router
next.

Also I need to figure out better calling semantics for the network
because it seems silly to have to cast to a void* style pthread function
pointer within the main function. It should be that I pass in the struct
and then interpret the void* as a pointer to the struct I'm expected.