Router function Unit Tests
md5sum a.out
a7ffb4868abce6acc937b630ce5086d3  a.out

Added testing for the router's function to determine the controller to
call. It's passing with flying colors now that I've redone some of the
logic in it. The urllength parameter should be removed since it's unused
entirely. I originally had it in place to make sure I didn't do anything
strange with overruning the buffers or losing parameters, but the min
was causing false positives for a lot of things. Not too mention that I
wasn't checking the return of the strncasecmp correctly... (for some
reason I forgot that string comparison functions return 0 on equal)

After fixing the router and making the main program, the server responds
with either 200 or 404 depending on if it recognizes the route or not.
This is a great step forward.