Added url query string parsing
md5sum a.out
9f934a61dc982f15673d0ca39eeda18a  a.out

Added strmap to the project. http://pokristensson.com/strmap.html It's a
great and simple to use string-only hashtable. It works well for parsing
query parameters.

Added and unit tested query parameter parsing. It appears to work
perfectly well, which is fantastic.

Now if it can handle JSON parsing as well we'll be all set. (although it
will be so simple)