HeartBeat Controller linked to Network
md5sum a.out
f1c94c2b7ab2c3cbd8ec89a65cbd9a5a  a.out

I've decided on a course of action for controllers. Each controller will
make a function titled: controllername_controller() that will take
whatever arguments are neccesary for it to function. For most this will
be the strmap of query parameters, the data sent with the request, and
the buffers to fill with output data.

The heartbeat controller is functional and working. Requesting "" "/"
"/api" "api/" will send back the heartbeat of the api. Any route that
matches a proper api route will give back a 200, and any route that
doesn't match will return a 404 like you would expect.