Added Comments Controller
md5sum a.out
7079c29d70f8c0026d0e68802f117e4e  a.out

Refactored and moved the parseURL function into the router code to make
the dependencies easier to manage for the controllers. The router also
defines a general error string formatter that takes a status code and
string message to return errors to the calling client.

The comment controller is able to accept and handle GET and POST
requests. The GET url parameters are parsed correctly and error messages
are returned if they query is invalid. No parsing of the POST request
has been done yet as I need a JSON validifier function first.