Added global variable for campaign id
md5sum a.out
428c80b24fa63535a9522efeea375b99  a.out

Added simple global variable. This is so that when you start the server
with the --scope or -s flag set so you can specify which campaign you
want to run as, all queries from henceforth (from the network layer at
least) will use the correct scope.