Full Stack Regular Comment Get.
md5sum a.out
f0a9e7c1ff36f9c3ee6e4bde42b23e8b  a.out

While the comment controller still needs some more work (including a
database migration) the basics are in place. GET requests to comments
are processed, the page is used, and type parsed for validity (although
the comment type is not used yet for a query, that's to come.)

There are a couple optimizations I can see happening. Like the strncat
going on to build the pages being exchanged for a custom string copying
function that uses the last known write position and continues from
there.