Fixed paging in comments db functions
md5sum a.out
cef06087a8ba7335bc37e122ea719d10  a.out

In order to do the nextUrl part of the returned JSON we always grab one
more then we need from the database. The existence of this struct
informs us that there are more results to be had.

Because of this, we need to subtract the page number from the limit.
This is because on the first query we'll grab one more then we need.
Then the next page we'll have offset by the proper amount minus the page
we're on. If you do the arithmetic/series out you'll see what I mean:

0,21,42,63, ... => (page-1)*21 (21 is the results per page +1)
0,20,40,60,80 .... => (page-1)*20

You can see pretty clearly that the actual objects returned from the
database advance by +page than what we give to the client.

This fix for the paging is only fixed on the comments functions in the
database. I  will fixup the other model queries once I begin their
controllers.