Added pin_id to comments
md5sum a.out
4d6577dc4fb48a978579d37616f14361  a.out

You'll need to migrate your database to the new version in order to run
this code. The README has been updated with the new sql as well as a
handy trigger. The triggers purpose is to update a comment's pin_id to
be the same as the owning pin if a marker gets inserted.
I've tested this and it works.

Also, I began refactoring the queries to use the sizeof operator
whenever possible to make it less of a pain to update the query sizes. I
need to go through and update them all to be safer in that way or risk
stack smashing.

Also, an idea that occured to me while trying to get the unit tests to
compile (the make tests command will not currently work) is to set up
some mocking behavior so that I don't need to link the entire
application together to test one thing. It's a pain to not be able to
test code without throwing a giant link statement into the make file. So
I might work on mocking some of the network interactions or something
soon. Also the global variable being shared poses a problem for testing
since it's not defined during the tests since it's defined in
green-serv.c