Fixed Tests. + Query size refactor
md5sum a.out
2a69924c4f815413df5f8bd428d671b8  a.out

Fixed the make tests command to work again. In order to use the shared
campaign id global you need to define the GREENSERV define at the _top_
of your file _before_ you include config.h otherwise it might be
declared an extern in your file and it won't have any idea where it's
coming from.

Refactor the database queries to use sizeof in as many places as
possible for the heap-allocated arrays. This cuts down on the use of
magic numbers and it just general more foolproof strategy.

I still want to fix up a mock-ed framework for the network tests and
whatnot. I think there would have to be a mock for each test that had
the neccesary functions prototypes and have functions that just returned
dummy values or some type of expected behavior from whatever interface
was being called.