Updated SQL tables for c_type
md5sum a.out
2a69924c4f815413df5f8bd428d671b8  a.out

Updated mySQL comments table to have a comment type field. Also added a
trigger to make sure that the values are always all uppercase within
that table.

This is just an update to the SQL, the code itself is not making use of
the comment types. Yet.