Type Filtering for Comments
md5sum a.out
2831e8db796d46bbdd2d6938b1dc5ee1  a.out

The comment controller GET requests can now handle the type= query
parameter being passed. A bit of work still needs to be done on the
uppercasing of the passed in result, but I think I can do that just on
the mySQL query level with UPPER and save myself the trouble of doing it
in the application itself.

The struct has been updated with a commentType field but no methods have
been created to set it safely or to display the type in the JSON. That
is the next step and some sanitization checks against the user input
will be the next bit to go. And then I will either move to handling the
POST request against the comments controller (and having a completely
functional comments controller) or will do the GET and DELETE request
handlers for the other API endpoints.